+++ core/server/api_container/server/module_store/module_launcher/module_launcher.go
@@ -14,11 +14,10 @@ import (
 	"github.com/kurtosis-tech/kurtosis-core/server/api_container/server/module_store/module_store_types"
 	"github.com/kurtosis-tech/kurtosis-core/server/commons"
 	"github.com/kurtosis-tech/kurtosis-core/server/commons/current_time_str_provider"
-	"github.com/kurtosis-tech/kurtosis-core/server/commons/object_labels_providers"
-	"github.com/kurtosis-tech/kurtosis-core/server/commons/object_name_providers"
 	"github.com/kurtosis-tech/kurtosis-module-api-lib/golang/kurtosis_module_docker_api"
 	"github.com/kurtosis-tech/kurtosis-module-api-lib/golang/kurtosis_module_rpc_api_bindings"
 	"github.com/kurtosis-tech/kurtosis-module-api-lib/golang/kurtosis_module_rpc_api_consts"
+	"github.com/kurtosis-tech/object-attributes-schema-lib/schema"
 	"github.com/kurtosis-tech/stacktrace"
 	"github.com/sirupsen/logrus"
 	"google.golang.org/grpc"
@@ -38,9 +37,7 @@ type ModuleLauncher struct {
 	// Modules have a connection to the API container, so the launcher must know about the API container's IP addr
 	apiContainerIpAddr string
 
-	enclaveObjNameProvider *object_name_providers.EnclaveObjectNameProvider
-
-	enclaveObjLabelsProvider *object_labels_providers.EnclaveObjectLabelsProvider
+	enclaveObjAttrsProvider schema.EnclaveObjectAttributesProvider
 
 	freeIpAddrTracker *commons.FreeIpAddrTracker
 
@@ -54,8 +51,8 @@ type ModuleLauncher struct {
 	enclaveDataDirpathOnHostMachine string
 }
 
-func NewModuleLauncher(dockerManager *docker_manager.DockerManager, apiContainerIpAddr string, enclaveObjNameProvider *object_name_providers.EnclaveObjectNameProvider, enclaveObjLabelsProvider *object_labels_providers.EnclaveObjectLabelsProvider, freeIpAddrTracker *commons.FreeIpAddrTracker, shouldPublishPorts bool, dockerNetworkId string, enclaveDataDirpathOnHostMachine string) *ModuleLauncher {
-	return &ModuleLauncher{dockerManager: dockerManager, apiContainerIpAddr: apiContainerIpAddr, enclaveObjNameProvider: enclaveObjNameProvider, enclaveObjLabelsProvider: enclaveObjLabelsProvider, freeIpAddrTracker: freeIpAddrTracker, shouldPublishPorts: shouldPublishPorts, dockerNetworkId: dockerNetworkId, enclaveDataDirpathOnHostMachine: enclaveDataDirpathOnHostMachine}
+func NewModuleLauncher(dockerManager *docker_manager.DockerManager, apiContainerIpAddr string, enclaveObjAttrsProvider schema.EnclaveObjectAttributesProvider, freeIpAddrTracker *commons.FreeIpAddrTracker, shouldPublishPorts bool, dockerNetworkId string, enclaveDataDirpathOnHostMachine string) *ModuleLauncher {
+	return &ModuleLauncher{dockerManager: dockerManager, apiContainerIpAddr: apiContainerIpAddr, enclaveObjAttrsProvider: enclaveObjAttrsProvider, freeIpAddrTracker: freeIpAddrTracker, shouldPublishPorts: shouldPublishPorts, dockerNetworkId: dockerNetworkId, enclaveDataDirpathOnHostMachine: enclaveDataDirpathOnHostMachine}
 }
 
 func (launcher ModuleLauncher) Launch(
@@ -100,8 +97,9 @@ func (launcher ModuleLauncher) Launch(
 	suffix := current_time_str_provider.GetCurrentTimeStr()
 	moduleGUID :=  module_store_types.ModuleGUID(string(moduleID) + "_" + suffix)
 
-	containerName := launcher.enclaveObjNameProvider.ForModuleContainer(moduleGUID)
-	containerLabels := launcher.enclaveObjLabelsProvider.ForModuleContainer(moduleGUID)
+	containerAttrs := launcher.enclaveObjAttrsProvider.ForModuleContainer(string(moduleGUID))
+	containerName := containerAttrs.GetName()
+	containerLabels := containerAttrs.GetLabels()
 	createAndStartArgs := docker_manager.NewCreateAndStartContainerArgsBuilder(
 		containerImage,
 		containerName,